+++ serialise/tests/Tests/Boundary.hs
@@ -10,22 +10,19 @@ module Tests.Boundary
   ( testTree -- :: TestTree
   ) where
 
-import           Control.DeepSeq
-import           Data.Bits
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as BSL
 import           Data.Either
 import           Data.Int
-import           Data.Monoid
 import           Data.Word
 import qualified Data.Text as T
-import qualified Data.Text.Encoding as T
 
 import           Codec.CBOR.Decoding
 import           Codec.CBOR.Encoding
 import           Codec.CBOR.Read
 import           Codec.CBOR.Write
 import           Codec.Serialise.Class
+import           Tests.Util
 
 import           Test.Tasty
 import           Test.Tasty.QuickCheck
@@ -33,26 +30,6 @@ import           Test.Tasty.QuickCheck
 #if !MIN_VERSION_base(4,8,0)
 import           Control.Applicative
 #endif
-
--- | Generate values of type 'a' embedded within (usually larger) type 'r' with
--- upped probabilities of getting neighbourhood of bounds of 'a'.
-arbitraryWithBounds
-  :: forall a r. (Bounded a, Integral a, Num r, Arbitrary r)
-  => a -> Gen r
-arbitraryWithBounds _ = frequency
-  [ (70, arbitrary)
-  -- Boundaries
-  , (5, pure $ fromIntegral (minBound     :: a))
-  , (5, pure $ fromIntegral (maxBound     :: a))
-  -- Near boundaries, in range
-  , (5, pure $ fromIntegral (minBound + 1 :: a))
-  , (5, pure $ fromIntegral (maxBound - 1 :: a))
-  -- Near boundaries, out of range (note: might overflow)
-  , (5, pure $ fromIntegral (minBound :: a) - 1)
-  , (5, pure $ fromIntegral (maxBound :: a) + 1)
-  ]
-
-----------------------------------------
 
 -- | CBOR can represent 64 bit negative and positive integers, hence we need
 -- wrapper for Integer to represent the whole range.
@@ -96,13 +73,14 @@ boundaryTest
                     Ord rep, Num rep, Serialise rep)
   => (forall s. Decoder s a)
   -> B a
-  -> Bool
+  -> Property
 boundaryTest dec a = if outsideRange
-                     then isLeft  a'
-                     else isRight a'
+                     then collect "outside" $ isLeft  a'
+                     else collect "inside"  $ isRight a'
   where
     a' = deserialiseFromBytes dec . toLazyByteString . encode $ unB a
 
+    -- Note that this is always true for a ~ rep.
     outsideRange = unB a < fromIntegral (minBound :: a)
                 || unB a > fromIntegral (maxBound :: a)
 
@@ -120,25 +98,16 @@ mkBoundaryTest aName dec decCan =
 
 ----------------------------------------
 
--- | Wrapper for list/map length.
-newtype Length = Length { unLength :: Word }
-
-instance Show Length where
-  showsPrec p = showsPrec p . unLength
-
-instance Arbitrary Length where
-  arbitrary = Length <$> arbitraryWithBounds (undefined::Int)
-
 -- | Check if deserialisation of map/list length deals properly with the ones
 -- out of range, i.e. fails to decode them.
 lenBoundaryTest
   :: (Word -> Encoding)
   -> (forall s. Decoder s Int)
   -> Length
-  -> Bool
+  -> Property
 lenBoundaryTest enc dec a = if outsideRange
-                            then isLeft  a'
-                            else isRight a'
+                            then collect "outside" $ isLeft  a'
+                            else collect "inside"  $ isRight a'
   where
     a' = deserialiseFromBytes dec . toLazyByteString . enc $ unLength a
 
@@ -156,22 +125,6 @@ mkLenBoundaryTest aName enc dec decCan =
   ]
 
 ----------------------------------------
-
--- | Generate CBOR prefix of non-empty string/bytes containing its length.
-mkLengthPrefix :: Bool -> Length -> BSL.ByteString
-mkLengthPrefix string (Length w)
-  | w <= 23         = BSL.pack $ [64 + stringBit + fromIntegral w]
-  | w <= 0xff       = BSL.pack $ [88 + stringBit] ++ f 1 w []
-  | w <= 0xffff     = BSL.pack $ [89 + stringBit] ++ f 2 w []
-  | w <= 0xffffffff = BSL.pack $ [90 + stringBit] ++ f 4 w []
-  | otherwise       = BSL.pack $ [91 + stringBit] ++ f 8 w []
-  where
-    stringBit :: Word8
-    stringBit = if string then 32 else 0
-
-    f :: Int -> Word -> [Word8] -> [Word8]
-    f 0 _ acc = acc
-    f k n acc = f (k - 1) (n `shiftR` 8) (fromIntegral n : acc)
 
 data StringLengthPrefix = StringLP Word BSL.ByteString
   deriving Show
@@ -203,25 +156,6 @@ stringBytesBoundaryTest =
 
 ----------------------------------------
 
--- | Wrapper for ByteString with Arbitrary instance that might produce a valid
--- UTF-8 encoding of a string.
-newtype MaybeText = MaybeText BS.ByteString
-  deriving Show
-instance Arbitrary MaybeText where
-  arbitrary = MaybeText . BS.pack <$> arbitrary
-
--- | Test that decoding of both valid and invalid CBOR strings produces output
--- without exceptions hidden within.
-utf8DecodingTest :: MaybeText -> Bool
-utf8DecodingTest (MaybeText bs) =
-  case (T.decodeUtf8' bs, deserialiseFromBytes decodeString s) of
-    v@(Right _, Right _) -> deepseq v True
-    v@(Left _,  Left _)  -> deepseq v True
-    _                    -> False
-  where
-    s = mkLengthPrefix True (Length . fromIntegral $ BS.length bs)
-     <> BSL.fromStrict bs
-
 testTree :: TestTree
 testTree = localOption (QuickCheckTests 1000) . testGroup "Boundary checks" $ concat
   [ mkBoundaryTest "Word"      decodeWord      decodeWordCanonical
@@ -239,9 +173,4 @@ testTree = localOption (QuickCheckTests 1000) . testGroup "Boundary checks" $ co
   , mkLenBoundaryTest "MapLen"  encodeMapLen  decodeMapLen  decodeMapLenCanonical
 
   , stringBytesBoundaryTest
-
-  -- TODO: move it somewhere else
-  , [testProperty "result of Text decoding doesn't contain unevaluated exceptions"
-                  utf8DecodingTest
-    ]
   ]
